# api/openapi.yaml
openapi: 3.1.0
info:
  title: Pantry API
  version: 0.1.0
  description: |
    API for Pantry. Responses and request bodies for item resources follow JSON:API media type
    (`application/vnd.api+json`).
servers:
  - url: http://localhost:4000
    description: Local development

tags:
  - name: Health
  - name: Items

paths:
  /healthz:
    get:
      tags: [Health]
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/HealthDocument'
              example:
                meta:
                  status: ok
  /api/items:
    get:
      tags: [Items]
      summary: List pantry items
      operationId: listItems
      parameters:
        - in: query
          name: sort
          required: false
          description: |
            JSON:API sort expression. Comma-separated fields are supported.
            Use `-` prefix for descending. Defaults to `id` (ascending).
            Example: `sort=best_before,-name`
          schema:
            type: string
            default: id
        - in: query
          name: q
          required: false
          description: |
            Case-insensitive search term matched against item name and comment.
            Example: `q=rice`
          schema:
            type: string
      responses:
        '200':
          description: List of items
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ItemListDocument'
              examples:
                twoItems:
                  summary: Two pantry items
                  value:
                    data:
                      - type: items
                        id: 019579e0-c393-7f0c-8f04-2ad9549f2772
                        attributes:
                          name: Rice
                          best_before: '2026-12-31'
                          content_amount: 500
                          content_unit: grams
                          packaging: bag
                          picture_key: items/rice.png
                          comment: Basmati
                          created_at: '2026-03-09T20:51:00Z'
                          updated_at: '2026-03-09T20:51:00Z'
                      - type: items
                        id: 019579e0-c39a-7a44-8d74-a8a1892d5b1d
                        attributes:
                          name: Milk
                          best_before: '2026-03-14'
                          content_amount: 1
                          content_unit: l
                          packaging: box
                          picture_key: items/milk.png
                          comment: null
                          created_at: '2026-03-09T20:52:00Z'
                          updated_at: '2026-03-09T20:52:00Z'
        '400':
          description: Invalid sort parameter
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '400'
                    title: invalid sorting
                    detail: sort must use fields id, name, best_before, created_at, updated_at (prefix with '-' for desc)
        '500':
          description: Internal error
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '500'
                    title: internal error
                    detail: database unavailable
  /api/uploads:
    post:
      tags: [Uploads]
      summary: Create a direct upload for an item picture
      operationId: createUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
            example:
              filename: rice.png
              content_type: image/png
      responses:
        '201':
          description: Presigned upload URL for direct S3 upload
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/UploadResponseDocument'
              example:
                data:
                  type: uploads
                  attributes:
                    picture_key: items/01957a1a-8c2f-7b1d-9b2f-2f5e7b9a0f91.png
                    upload_url: https://s3.example.com/pantry/items/01957a1a-8c2f-7b1d-9b2f-2f5e7b9a0f91.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=...
                    headers:
                      Content-Type: image/png
        '400':
          description: Invalid upload request
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '400'
                    title: invalid content_type
                    detail: only jpeg, png, webp, or gif images are supported
        '415':
          description: Unsupported media type
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '415'
                    title: unsupported media type
                    detail: content-type must be application/json
        '500':
          description: Internal error
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '500'
                    title: upload error
                    detail: could not create upload URL
  /api/uploads/preview:
    get:
      tags: [Uploads]
      summary: Create a presigned preview URL for an item picture
      operationId: getUploadPreview
      parameters:
        - in: query
          name: picture_key
          required: true
          schema:
            type: string
          description: The picture key stored on the item
      responses:
        '200':
          description: Presigned preview URL
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/UploadPreviewDocument'
              example:
                data:
                  type: uploads
                  attributes:
                    picture_key: items/01957a1a-8c2f-7b1d-9b2f-2f5e7b9a0f91.png
                    preview_url: https://s3.example.com/pantry/items/01957a1a-8c2f-7b1d-9b2f-2f5e7b9a0f91.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=...
        '400':
          description: Invalid picture key
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '400'
                    title: invalid picture_key
                    detail: picture_key query parameter is required
        '500':
          description: Internal error
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '500'
                    title: preview error
                    detail: could not create preview URL
    post:
      tags: [Items]
      summary: Create a pantry item
      operationId: createItem
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/CreateItemDocument'
            examples:
              rice:
                summary: Create rice item
                value:
                  data:
                    type: items
                    attributes:
                      name: Rice
                      best_before: '2026-12-31'
                      content_amount: 500
                      content_unit: grams
                      packaging: bag
                      picture_key: items/rice.png
                      comment: Basmati
      responses:
        '201':
          description: Item created
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ItemSingleDocument'
              example:
                data:
                  type: items
                  id: 019579e0-c3b5-7424-bf89-ae567634a2ff
                  attributes:
                    name: Rice
                    best_before: '2026-12-31'
                    content_amount: 500
                    content_unit: grams
                    packaging: bag
                    picture_key: items/rice.png
                    comment: Basmati
                    created_at: '2026-03-09T20:55:00Z'
                    updated_at: '2026-03-09T20:55:00Z'
        '400':
          description: Invalid request payload or validation error
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              examples:
                invalidDate:
                  value:
                    errors:
                      - status: '400'
                        title: invalid best_before
                        detail: best_before must use YYYY-MM-DD
                validation:
                  value:
                    errors:
                      - status: '400'
                        title: validation error
                        detail: name is required
        '409':
          description: Invalid JSON:API resource type
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '409'
                    title: invalid type
                    detail: data.type must be items
        '415':
          description: Unsupported media type
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '415'
                    title: unsupported media type
                    detail: content-type must be application/vnd.api+json
  /api/items/{id}:
    patch:
      tags: [Items]
      summary: Update a pantry item
      operationId: updateItem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/UpdateItemDocument'
      responses:
        '200':
          description: Item updated
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ItemSingleDocument'
        '400':
          description: Invalid request payload or validation error
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
        '404':
          description: Item not found
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
        '409':
          description: Invalid JSON:API resource type or id mismatch
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
        '415':
          description: Unsupported media type
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
    delete:
      tags: [Items]
      summary: Delete a pantry item
      operationId: deleteItem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deleted
        '404':
          description: Item not found
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'
              example:
                errors:
                  - status: '404'
                    title: not found
                    detail: item not found
        '500':
          description: Internal error
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorDocument'

components:
  schemas:
    HealthDocument:
      type: object
      additionalProperties: false
      example:
        meta:
          status: ok
      properties:
        meta:
          type: object
          additionalProperties: false
          properties:
            status:
              type: string
              enum: [ok]
          required: [status]
      required: [meta]

    ItemResource:
      type: object
      additionalProperties: false
      example:
        type: items
        id: 019579e0-c393-7f0c-8f04-2ad9549f2772
        attributes:
          name: Rice
          best_before: '2026-12-31'
          content_amount: 500
          content_unit: grams
          packaging: bag
          picture_key: items/rice.png
          comment: Basmati
          created_at: '2026-03-09T20:51:00Z'
          updated_at: '2026-03-09T20:51:00Z'
      properties:
        type:
          type: string
          enum: [items]
        id:
          type: string
          description: UUIDv7 identifier
          format: uuid
        attributes:
          $ref: '#/components/schemas/ItemAttributes'
      required: [type, id, attributes]

    ItemAttributes:
      type: object
      additionalProperties: false
      example:
        name: Rice
        best_before: '2026-12-31'
        content_amount: 500
        content_unit: grams
        packaging: bag
        picture_key: items/rice.png
        comment: Basmati
        created_at: '2026-03-09T20:51:00Z'
        updated_at: '2026-03-09T20:51:00Z'
      properties:
        name:
          type: string
        best_before:
          type: string
          format: date
        content_amount:
          type: number
          exclusiveMinimum: 0
        content_unit:
          type: string
          enum: [grams, ml, l]
        packaging:
          type: string
          enum: [can, box, bag, jar, other]
        picture_key:
          type:
            - string
            - 'null'
        comment:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
        - best_before
        - content_amount
        - content_unit
        - packaging
        - created_at
        - updated_at

    UploadRequest:
      type: object
      additionalProperties: false
      properties:
        filename:
          type: string
        content_type:
          type: string
      required: [filename, content_type]

    UploadAttributes:
      type: object
      additionalProperties: false
      properties:
        picture_key:
          type: string
        upload_url:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string
      required: [picture_key, upload_url, headers]

    UploadPreviewAttributes:
      type: object
      additionalProperties: false
      properties:
        picture_key:
          type: string
        preview_url:
          type: string
          format: uri
      required: [picture_key, preview_url]

    UploadResponseDocument:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            type:
              type: string
              enum: [uploads]
            attributes:
              $ref: '#/components/schemas/UploadAttributes'
          required: [type, attributes]
      required: [data]

    UploadPreviewDocument:
      type: object
      additionalProperties: false
      properties:
        data:
          type: object
          additionalProperties: false
          properties:
            type:
              type: string
              enum: [uploads]
            attributes:
              $ref: '#/components/schemas/UploadPreviewAttributes'
          required: [type, attributes]
      required: [data]

    ItemListDocument:
      type: object
      additionalProperties: false
      example:
        data:
          - type: items
            id: 019579e0-c393-7f0c-8f04-2ad9549f2772
            attributes:
              name: Rice
              best_before: '2026-12-31'
              content_amount: 500
              content_unit: grams
              packaging: bag
              picture_key: items/rice.png
              comment: Basmati
              created_at: '2026-03-09T20:51:00Z'
              updated_at: '2026-03-09T20:51:00Z'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ItemResource'
      required: [data]

    ItemSingleDocument:
      type: object
      additionalProperties: false
      example:
        data:
          type: items
          id: 019579e0-c3b5-7424-bf89-ae567634a2ff
          attributes:
            name: Rice
            best_before: '2026-12-31'
            content_amount: 500
            content_unit: grams
            packaging: bag
            picture_key: items/rice.png
            comment: Basmati
            created_at: '2026-03-09T20:55:00Z'
            updated_at: '2026-03-09T20:55:00Z'
      properties:
        data:
          $ref: '#/components/schemas/ItemResource'
      required: [data]

    CreateItemDocument:
      type: object
      additionalProperties: false
      example:
        data:
          type: items
          attributes:
            name: Rice
            best_before: '2026-12-31'
            content_amount: 500
            content_unit: grams
            packaging: bag
            picture_key: items/rice.png
            comment: Basmati
      properties:
        data:
          $ref: '#/components/schemas/CreateItemResource'
      required: [data]

    CreateItemResource:
      type: object
      additionalProperties: false
      example:
        type: items
        attributes:
          name: Rice
          best_before: '2026-12-31'
          content_amount: 500
          content_unit: grams
          packaging: bag
          picture_key: items/rice.png
          comment: Basmati
      properties:
        type:
          type: string
          enum: [items]
        attributes:
          $ref: '#/components/schemas/CreateItemAttributes'
      required: [type, attributes]

    CreateItemAttributes:
      type: object
      additionalProperties: false
      example:
        name: Rice
        best_before: '2026-12-31'
        content_amount: 500
        content_unit: grams
        packaging: bag
        picture_key: items/rice.png
        comment: Basmati
      properties:
        name:
          type: string
          minLength: 1
        best_before:
          type: string
          format: date
        content_amount:
          type: number
          exclusiveMinimum: 0
        content_unit:
          type: string
          enum: [grams, ml, l]
        packaging:
          type: string
          enum: [can, box, bag, jar, other]
        picture_key:
          type:
            - string
            - 'null'
        comment:
          type:
            - string
            - 'null'
      required:
        - name
        - best_before
        - content_amount
        - content_unit
        - packaging

    UpdateItemDocument:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/UpdateItemResource'
      required: [data]

    UpdateItemResource:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [items]
        id:
          type: string
          format: uuid
        attributes:
          $ref: '#/components/schemas/CreateItemAttributes'
      required: [type, id, attributes]

    ErrorDocument:
      type: object
      additionalProperties: false
      example:
        errors:
          - status: '400'
            title: validation error
            detail: name is required
      properties:
        errors:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ErrorObject'
      required: [errors]

    ErrorObject:
      type: object
      additionalProperties: false
      example:
        status: '400'
        title: validation error
        detail: name is required
      properties:
        status:
          type: string
        title:
          type: string
        detail:
          type: string
      required: [status, title, detail]